refactor(console): generic and impl for account
